# Optimize JsonWriter APIs by using the TryWrite approach