--- a/src/System.Text.JsonLab/System/Text/Json/JsonWriterHelper.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonWriterHelper.cs
@@ -103,5 +103,86 @@ namespace System.Text.JsonLab
 
             return digits;
         }
+
+        // Borrowed from https://github.com/dotnet/corefx/blob/master/src/System.Memory/src/System/Buffers/Text/Utf8Formatter/Utf8Formatter.Integer.Signed.Default.cs#L16
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryFormatInt64Default(long value, Span<byte> destination, out int bytesWritten)
+        {
+            if ((ulong)value < 10)
+            {
+                return TryFormatUInt32SingleDigit((uint)value, destination, out bytesWritten);
+            }
+
+            return TryFormatInt64MultipleDigits(value, destination, out bytesWritten);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static bool TryFormatUInt32SingleDigit(uint value, Span<byte> destination, out int bytesWritten)
+        {
+            if (destination.Length == 0)
+            {
+                bytesWritten = 0;
+                return false;
+            }
+            destination[0] = (byte)('0' + value);
+            bytesWritten = 1;
+            return true;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static bool TryFormatInt64MultipleDigits(long value, Span<byte> destination, out int bytesWritten)
+        {
+            if (value < 0)
+            {
+                value = -value;
+                int digitCount = CountDigits((ulong)value);
+                // WriteDigits does not do bounds checks
+                if (digitCount >= destination.Length)
+                {
+                    bytesWritten = 0;
+                    return false;
+                }
+                destination[0] = (byte)'-';
+                bytesWritten = digitCount + 1;
+                WriteDigits((ulong)value, destination.Slice(1, digitCount));
+                return true;
+            }
+            else
+            {
+                return TryFormatUInt64MultipleDigits((ulong)value, destination, out bytesWritten);
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static bool TryFormatUInt64MultipleDigits(ulong value, Span<byte> destination, out int bytesWritten)
+        {
+            int digitCount = CountDigits(value);
+            // WriteDigits does not do bounds checks
+            if (digitCount > destination.Length)
+            {
+                bytesWritten = 0;
+                return false;
+            }
+            bytesWritten = digitCount;
+            WriteDigits(value, destination.Slice(0, digitCount));
+            return true;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void WriteDigits(ulong value, Span<byte> buffer)
+        {
+            // We can mutate the 'value' parameter since it's a copy-by-value local.
+            // It'll be used to represent the value left over after each division by 10.
+
+            for (int i = buffer.Length - 1; i >= 1; i--)
+            {
+                ulong temp = '0' + value;
+                value /= 10;
+                buffer[i] = (byte)(temp - (value * 10));
+            }
+
+            Debug.Assert(value < 10);
+            buffer[0] = (byte)('0' + value);
+        }
     }
 }