``` ini

BenchmarkDotNet=v0.10.14.534-nightly, OS=Windows 10.0.14393.2312 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=3328126 Hz, Resolution=300.4694 ns, Timer=TSC
.NET Core SDK=2.1.301
  [Host]     : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.1 (CoreCLR 4.6.26606.02, CoreFX 4.6.26606.05), 64bit RyuJIT


```
**BEFORE (master):**

|                         Method | Formatted | size |         Mean |       Error |      StdDev | Allocated |
|------------------------------- |---------- |----- |-------------:|------------:|------------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |     **72.55 ns** |   **0.8945 ns** |   **0.7469 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |    **160.18 ns** |   **0.6331 ns** |   **0.5286 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** | **12,708.05 ns** |  **76.4989 ns** |  **71.5571 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |    **769.99 ns** |  **15.4300 ns** |  **27.0244 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |     93.48 ns |   1.9026 ns |   3.3819 ns |       0 B |
|  **WriterSystemTextJsonArrayOnly** |      **True** |    **1** |     **86.39 ns** |   **1.0231 ns** |   **0.9570 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** |   **10** |    **209.60 ns** |   **2.9768 ns** |   **2.7845 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** | **1000** | **16,923.45 ns** | **303.0954 ns** | **283.5156 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |      **True** |    **?** |    **898.10 ns** |  **10.5260 ns** |   **9.8460 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |      True |    ? |    104.96 ns |   2.4575 ns |   5.5970 ns |       0 B |

**AFTER (this PR):**

|                         Method | Formatted | size |         Mean |       Error |        StdDev | Allocated |
|------------------------------- |---------- |----- |-------------:|------------:|--------------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |     **53.17 ns** |   **1.0753 ns** |     **0.8979 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |    **124.29 ns** |   **0.6007 ns** |     **0.5325 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** |  **9,874.39 ns** |  **65.1725 ns** |    **60.9623 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |  **1,061.03 ns** |  **21.1885 ns** |    **38.2072 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |     69.20 ns |   1.4188 ns |     2.8006 ns |       0 B |
|  **WriterSystemTextJsonArrayOnly** |      **True** |    **1** |     **84.37 ns** |   **1.7058 ns** |     **2.3913 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** |   **10** |    **225.26 ns** |   **4.5097 ns** |     **9.1098 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |      **True** | **1000** | **18,291.06 ns** | **531.5408 ns** | **1,558.9170 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |      **True** |    **?** |    **908.00 ns** |  **18.0899 ns** |    **26.5159 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |      True |    ? |    106.37 ns |   1.5931 ns |     1.4122 ns |       0 B |

cc @benaadams, @GrabYourPitchforks