@ahsonkhan 2018-06-29
I am not sure what is going on with the Basic test. I seem to be hitting some sort of **performance cliff** and I can't figure out why. The total bytes written is ~223 (well below the buffer size, so it doesn't seem to be due to re-sizing/etc.).

Writing the 10 value array requires ~130 ns (as seen in the WriterSystemTextJsonArrayOnly test). At some point, writing some more ends up tanking performance. I am not sure where the bug is. **Any ideas?**

I expect the benchmark to take ~600 ns given the results of writing more and more json.


|                     Method | Formatted |     Mean |    Error |   StdDev | Allocated |
|--------------------------- |---------- |---------:|---------:|---------:|----------:|
| WriterSystemTextJsonBasic1 |     False | 114.3 ns | 2.283 ns | 2.718 ns |       0 B |
| WriterSystemTextJsonBasic2 |     False | 215.5 ns | 3.493 ns | 2.917 ns |       0 B |
| WriterSystemTextJsonBasic3 |     False | 469.4 ns | 7.044 ns | 6.589 ns |       0 B |
| WriterSystemTextJsonBasic4 |     False | 562.3 ns | 7.480 ns | 6.246 ns |       0 B |
| WriterSystemTextJsonBasic5 |     False | 974.5 ns | 3.736 ns | 2.916 ns |       0 B |


If I change the benchmark to the following (where I move the array writing to a separate method), I get the ~600 ns result:
```C#
private static void WriterSystemTextJsonBasicUtf8_5(bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
{
    var json = new JsonWriter<ArrayFormatterWrapper>(output, formatted);

    json.WriteObjectStart();
    json.WriteAttribute("age", 42);
    json.WriteAttribute("first", "John");
    json.WriteAttribute("last", "Smith");
    json.WriteArrayStart("phoneNumbers");
    json.WriteValue("425-000-1212");
    json.WriteValue("425-000-1213");
    json.WriteArrayEnd();
    json.WriteObjectStart("address");
    json.WriteAttribute("street", "1 Microsoft Way");
    json.WriteAttribute("city", "Redmond");
    json.WriteAttribute("zip", 98052);
    json.WriteObjectEnd();

    WriterSystemTextJsonArrayOnlyUtf8(ref json, formatted, output, data);

    json.WriteObjectEnd();
    json.Flush();
}

private static void WriterSystemTextJsonArrayOnlyUtf8(ref JsonWriter<ArrayFormatterWrapper> json, bool formatted, ArrayFormatterWrapper output, ReadOnlySpan<int> data)
{
    json.WriteArrayStart("ExtraArray");
    for (var i = 0; i < data.Length; i++)
    {
        json.WriteValue(data[i]);
    }
    json.WriteArrayEnd();
    json.Flush();
}
```