@benaadams 2018-06-29
Random thought, what's the asm size for the method? If the inlines might cause it to cross a page (e.g. > 4096 bytes) might be an issue for the instruction cache?