@ahsonkhan 2018-06-29
Looks like a lot more struct copies are happening in the "slow" case (where all the json writing is within a single method):

![image](https://user-images.githubusercontent.com/6527137/42075996-256e3108-7b28-11e8-8789-97859931621f.png)

cc @AndyAyersMS, any ideas why?

**The entire disassembly is available here:**
[results-slow.zip](https://github.com/dotnet/corefxlab/files/2147962/results-slow.zip)
[results-fast.zip](https://github.com/dotnet/corefxlab/files/2147963/results-fast.zip)