@ahsonkhan 2018-06-29
> Random thought, what's the asm size for the method? If the inlines might cause it to cross a page (e.g. > 4096 bytes) might be an issue for the instruction cache?'

**Fast:** 
`System.Text.JsonLab.Benchmarks.JsonWriterPerf.WriterSystemTextJsonBasicUtf8_5(Boolean, System.Text.Formatting.ArrayFormatterWrapper, System.ReadOnlySpan)` starts on address `0xf64f2001` and ends on `0xf64f371f`, i.e. **5918 bytes**

**Slow:**
`System.Text.JsonLab.Benchmarks.JsonWriterPerf.WriterSystemTextJsonBasicUtf8_5(Boolean, System.Text.Formatting.ArrayFormatterWrapper, System.ReadOnlySpan)` starts on address `0xf6511fff` and ends on `0xf651456c`, i.e. **9581 bytes**

Although, I don't see the call to `WriterSystemTextJsonArrayOnlyUtf8(ref...`) in the disassembly for the fast case.