@benaadams 2018-06-29
Probably more related to the struct copies then since they both go over?