@Tornhoof 2018-06-29
I had weird struct copies in some cases in Spanjson. Added a few in modifiers in front of Readonlyspan helped.