@ahsonkhan 2018-06-30
**By being less aggressive about aggressive inlining,** we help the `Basic` case (now its perf is as expected, 70% reduction) and generally don't hurt performance in other cases too much (small json like hello world regresses ~10%).

cc @jkotas 

**BEFORE:**

|                         Method | Formatted | size |         Mean |      Error |     StdDev | Allocated |
|------------------------------- |---------- |----- |-------------:|-----------:|-----------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |     **53.97 ns** |  **0.1883 ns** |  **0.1572 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |    **123.44 ns** |  **0.2069 ns** |  **0.1834 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** | **10,160.26 ns** | **26.1318 ns** | **23.1651 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |    **940.65 ns** |  **2.0003 ns** |  **1.7732 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |     61.14 ns |  0.1754 ns |  0.1555 ns |       0 B |

**AFTER:**

|                         Method | Formatted | size |        Mean |      Error |     StdDev | Allocated |
|------------------------------- |---------- |----- |------------:|-----------:|-----------:|----------:|
|  **WriterSystemTextJsonArrayOnly** |     **False** |    **1** |    **63.01 ns** |  **0.2694 ns** |  **0.2388 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** |   **10** |   **135.17 ns** |  **0.6625 ns** |  **0.6197 ns** |       **0 B** |
|  **WriterSystemTextJsonArrayOnly** |     **False** | **1000** | **9,945.08 ns** | **37.4193 ns** | **33.1712 ns** |       **0 B** |
|      **WriterSystemTextJsonBasic** |     **False** |    **?** |   **529.19 ns** |  **2.2597 ns** |  **2.1138 ns** |       **0 B** |
| WriterSystemTextJsonHelloWorld |     False |    ? |    65.68 ns |  0.2237 ns |  0.1868 ns |       0 B |